Fix app loaders to resolve from bundle root
